updateSQL Function added
Non-functional, but all of the framework and some basic things are there. It does attempt to INSERT but it does not appear to fully insert into the SQLite DB for some reason.
No other errors noted, as the edit function has not been implemented yet. Will implement edit and then return to troubleshooting.