fixed database save location
changed the behavior so that it pulls the directory of the python script being used, rather than the terminal location - this was causing discrepencies between running the file in terminal from differing directories and when running the script in the folder. Basically - it always makes the folder at folder_where_2dopy_is/data vs wherever the script was run from.